reframe nft configs
